pyproject.toml
[tool.poetry]
name = "frog"
version = "0.1.0"
description = "Host-hopping infrastructure configuration tool built with Mitogen"
authors = ["Sean Johnson <me@seanj.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
mitogen = "^0.3.2"
Jinja2 = "^3.1.2"
testinfra = "^6.0.0"
PyYAML = "^6.0"
click = "^8.1.3"
hvac = "^0.11.2"
types-PyYAML = "^6.0.7"
netaddr = "^0.8.0"
texttable = "^1.6.4"
better-exceptions = "^0.3.3"
psutil = "^5.9.1"
netifaces = "^0.11.0"
regex = "^2022.4.24"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.2"
pytest-testinfra = "^6.7.0"
pytest-cov = "^3.0.0"
mypy = "^0.950"
flake8 = "^4.0.1"

[tool.black]
line-length = 88
target-version = ["py37", "py38"]
include = ".pyi?$"

[tool.poetry.scripts]
frog = "frog.cli:root"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
